+++ resolwe_bio/tests/processes/test_expression_aggregator.py
@@ -1,4 +1,5 @@
 # pylint: disable=missing-docstring
+from resolwe.flow.models import Data
 from resolwe.test import tag_process
 from resolwe_bio.utils.test import BioProcessTestCase
 
@@ -7,26 +8,35 @@ class ExpressionAggregatorTestCase(BioProcessTestCase):
     @tag_process('expression-aggregator')
     def test_expression_aggregator(self):
         with self.preparation_stage():
-            descriptor_hs = {'sample': {'organism': 'Homo sapiens'}}
+            descriptor_ery = {'sample': {'cell_type': 'Erythrocyte'}}
             expression1 = self.prepare_expression(f_rc='exp_1_rc.tab.gz', f_exp='exp_1_tpm.tab.gz',
-                                                  f_type='TPM', descriptor=descriptor_hs,
+                                                  f_type='TPM', descriptor=descriptor_ery,
                                                   species='Homo sapiens', build='ens_90')
+
+            descriptor_mono = {'sample': {'cell_type': 'Monocyte'}}
             expression2 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz',
-                                                  f_type='TPM', descriptor=descriptor_hs,
+                                                  f_type='TPM', descriptor=descriptor_mono,
                                                   species='Homo sapiens', build='ens_90')
 
-            descriptor_mm = {'sample': {'organism': 'Mus musculus'}}
             expression3 = self.prepare_expression(f_rc='exp_1_rc.tab.gz', f_exp='exp_1_tpm.tab.gz',
-                                                  f_type='TPM', descriptor=descriptor_mm,
-                                                  species='Mus musculus', build='ens_90')
-            expression4 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz',
-                                                  f_type='TPM', descriptor=descriptor_mm,
+                                                  f_type='TPM', descriptor=descriptor_ery,
                                                   species='Mus musculus', build='ens_90')
 
+            expression4 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz',
+                                                  f_type='TPM', descriptor=descriptor_mono,
+                                                  species='Mus musculus', build='ens_90')
+
+        # Expect the process to fail if expression data from multiple species is used
         inputs = {
-            'exps': [expression1.id, expression2.id, expression3.id, expression4.id],
-            'group_by': 'sample.organism'
+            'exps': [expression1.id, expression2.id, expression3.id],
+            'group_by': 'sample.cell_type',
         }
+        expression_aggregator = self.run_process('expression-aggregator', inputs, Data.STATUS_ERROR)
+
+        inputs['exps'] = [expression3.id, expression4.id]
+        expression_aggregator_mm = self.run_process('expression-aggregator', inputs)
+
+        inputs['exps'] = [expression1.id, expression2.id]
         expression_aggregator = self.run_process('expression-aggregator', inputs)
 
         saved_matrix, test_matrix = self.get_json('exp_matrix.json.gz', expression_aggregator.output['exp_matrix'])
@@ -39,3 +49,8 @@ class ExpressionAggregatorTestCase(BioProcessTestCase):
         )
         self.assertFields(expression_aggregator, 'source', 'DICTYBASE')
         self.assertFields(expression_aggregator, 'exp_type', 'TPM')
+        self.assertFields(expression_aggregator, 'species', 'Homo sapiens')
+
+        # Do not allow appending expression data of one species to the existing Aggregator objects of different species
+        inputs['expr_aggregator'] = expression_aggregator_mm.id
+        expression_aggregator = self.run_process('expression-aggregator', inputs, Data.STATUS_ERROR)